+++ akshare/fx/currency_investing.py
@@ -14,7 +14,7 @@ import requests
 from bs4 import BeautifulSoup
 from tqdm import tqdm
 
-from akshare.index.cons import short_headers, long_headers
+from akshare.index.cons import short_headers
 
 
 def _currency_name_url() -> dict:
@@ -92,8 +92,21 @@ def currency_hist(
         "add-missing-rows": "false",
     }
     headers = {
+        "accept": "application/json, text/plain, */*",
+        "accept-encoding": "gzip, deflate, br",
+        "accept-language": "zh-CN,zh;q=0.9,en;q=0.8",
+        "cache-control": "no-cache",
         "domain-id": "cn",
-        "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.0.0 Safari/537.36",
+        "origin": "https://cn.investing.com",
+        "pragma": "no-cache",
+        "referer": "https://cn.investing.com/",
+        "sec-ch-ua": '"Google Chrome";v="105", "Not)A;Brand";v="8", "Chromium";v="105"',
+        "sec-ch-ua-mobile": '"?0"',
+        "sec-ch-ua-platform": '"Windows"',
+        "sec-fetch-dest": "empty",
+        "sec-fetch-mode": "cors",
+        "sec-fetch-site": "same-site",
+        "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/105.0.0.0 Safari/537.36",
     }
     r = requests.get(url, params=params, headers=headers)
     data_json = r.json()