+++ akshare/stock_feature/stock_lhb_em.py
@@ -177,7 +177,7 @@ def stock_lhb_stock_statistic_em(symbol: str = "近一月") -> pd.DataFrame:
 
 
 def stock_lhb_jgmmtj_em(
-    start_date: str = "20220311", end_date: str = "20220315"
+    start_date: str = "20220906", end_date: str = "20220906"
 ) -> pd.DataFrame:
     """
     东方财富网-数据中心-龙虎榜单-机构买卖每日统计
@@ -224,13 +224,15 @@ def stock_lhb_jgmmtj_em(
         "市场总成交额",
         "机构净买额占总成交额比",
         "换手率",
-        "-",
+        "流通市值",
         "上榜原因",
-        "上榜后1日",
-        "上榜后2日",
         "-",
-        "上榜后5日",
-        "上榜后10日",
+        "-",
+        "-",
+        "-",
+        "-",
+        "-",
+        "-",
         "-",
         "-",
     ]
@@ -239,7 +241,6 @@ def stock_lhb_jgmmtj_em(
             "序号",
             "代码",
             "名称",
-            "上榜日期",
             "收盘价",
             "涨跌幅",
             "买方机构数",
@@ -250,14 +251,24 @@ def stock_lhb_jgmmtj_em(
             "市场总成交额",
             "机构净买额占总成交额比",
             "换手率",
+            "流通市值",
             "上榜原因",
-            "上榜后1日",
-            "上榜后2日",
-            "上榜后5日",
-            "上榜后10日",
+            "上榜日期",
         ]
     ]
     temp_df["上榜日期"] = pd.to_datetime(temp_df["上榜日期"]).dt.date
+    temp_df['收盘价'] = pd.to_numeric(temp_df['收盘价'], errors="coerce")
+    temp_df['涨跌幅'] = pd.to_numeric(temp_df['涨跌幅'], errors="coerce")
+    temp_df['买方机构数'] = pd.to_numeric(temp_df['买方机构数'], errors="coerce")
+    temp_df['卖方机构数'] = pd.to_numeric(temp_df['卖方机构数'], errors="coerce")
+    temp_df['机构买入总额'] = pd.to_numeric(temp_df['机构买入总额'], errors="coerce")
+    temp_df['机构卖出总额'] = pd.to_numeric(temp_df['机构卖出总额'], errors="coerce")
+    temp_df['机构买入净额'] = pd.to_numeric(temp_df['机构买入净额'], errors="coerce")
+    temp_df['市场总成交额'] = pd.to_numeric(temp_df['市场总成交额'], errors="coerce")
+    temp_df['机构净买额占总成交额比'] = pd.to_numeric(temp_df['机构净买额占总成交额比'], errors="coerce")
+    temp_df['换手率'] = pd.to_numeric(temp_df['换手率'], errors="coerce")
+    temp_df['流通市值'] = pd.to_numeric(temp_df['流通市值'], errors="coerce")
+
     return temp_df
 
 
@@ -533,7 +544,7 @@ if __name__ == "__main__":
     print(stock_lhb_stock_statistic_em_df)
 
     stock_lhb_jgmmtj_em_df = stock_lhb_jgmmtj_em(
-        start_date="20220311", end_date="20220315"
+        start_date="20220904", end_date="20220906"
     )
     print(stock_lhb_jgmmtj_em_df)